blue: Initial commit with basic lexer and tests
I added spans so that the printing can be nicer.
There is 1 test case that covers it at the moment.

Still need to think about token that are necessary and not necessary.

Any new tokens added should have the tests added for them as well.

Make is used for convenience but is not really necessary.

The flag package will probably be replaced by a better cli at some point

There are still TODOs left with regards to refactoring but the code has
been separated by visibility.